+++ app/lib/service/entrypoint/search.dart
@@ -7,6 +7,7 @@ import 'dart:isolate';
 
 import 'package:args/command_runner.dart';
 import 'package:logging/logging.dart';
+import 'package:pub_dev/search/backend.dart';
 
 import '../../search/dart_sdk_mem_index.dart';
 import '../../search/flutter_sdk_mem_index.dart';
@@ -37,8 +38,9 @@ class SearchCommand extends Command {
     await startIsolates(
       logger: _logger,
       frontendEntryPoint: _main,
+      workerEntryPoint: _worker,
       frontendCount: 1,
-      workerCount: 0,
+      workerCount: 1,
     );
   }
 }
@@ -50,7 +52,6 @@ Future _main(FrontendEntryMessage message) async {
     statsConsumerPort: statsConsumer.sendPort,
   ));
 
-  setupSearchPeriodicTasks();
   await popularityStorage.start();
 
   final ReceivePort taskReceivePort = ReceivePort();
@@ -71,4 +72,11 @@ Future _main(FrontendEntryMessage message) async {
   });
 
   await runHandler(_logger, searchServiceHandler);
+}
+
+Future _worker(WorkerEntryMessage message) async {
+  message.protocolSendPort.send(WorkerProtocolMessage());
+  await popularityStorage.start();
+  setupSearchPeriodicTasks();
+  await searchBackend.updateSnapshotInForeverLoop();
 }