+++ app/test/search/handlers_test.dart
@@ -12,6 +12,7 @@ import 'package:pub_dev/search/flutter_sdk_mem_index.dart';
 import 'package:pub_dev/search/mem_index.dart';
 import 'package:pub_dev/search/search_service.dart';
 import 'package:pub_dev/shared/exceptions.dart';
+import 'package:pub_dev/task/global_lock.dart';
 import 'package:test/test.dart';
 
 import '../frontend/handlers/_utils.dart' as _utils;
@@ -195,6 +196,17 @@ class MockSearchBackend implements SearchBackend {
 
   @override
   Future<void> close() {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<Never> updateSnapshotInForeverLoop() {
+    throw UnimplementedError();
+  }
+
+  @override
+  Future<void> doCreateAndUpdateSnapshot(GlobalLockClaim claim,
+      {Duration sleepDuration = const Duration(minutes: 2)}) {
     throw UnimplementedError();
   }
 }